se implemento algo de dpf

+++ lib/modules/bank_guarantees/bank_security_solicitation_result/domain/entities/bank_security_solicitation_result_entity.dart
@@ -20,7 +20,7 @@ class BankSecuritySolicitationResultEntity {
   final String beneficiario;
   final String solicitante;
   final String documentoOrdenante;
-  final int comisionFianzaBancaria;
+  final double comisionFianzaBancaria;
 
   BankSecuritySolicitationResultEntity({
     required this.codigoSolicitud,
